Fix clang build
Summary:
https://ci-builds.fb.com/view/rocksdb/job/rocksdb_clang_build/893/console
Fixing clang build

Test Plan:
make clean
USE_CLANG=1 make all -j64

Reviewers: sdong, noetzli, yhchiang, igor

Reviewed By: igor

Subscribers: dhruba

Differential Revision: https://reviews.facebook.net/D43959

+++ db/compaction_job_test.cc
@@ -230,7 +230,7 @@ class CompactionJobTest : public testing::Test {
     for (size_t level = 0; level < input_files.size(); level++) {
       auto level_files = input_files[level];
       CompactionInputFiles compaction_level;
-      compaction_level.level = level;
+      compaction_level.level = static_cast<int>(level);
       compaction_level.files.insert(compaction_level.files.end(),
           level_files.begin(), level_files.end());
       compaction_input_files.push_back(compaction_level);
